Define the initial OpenAPI specification
Define the OpenAPI specification with endpoints to store data, retrieve
data and check the status of a data by ID.

The API specification uses the prefix `/v0` while it is being developed.
All bets are off in this version in terms of breaking changes.

The API introduces `/data` namespace to encapsulate data management and
allow room for other domain-specific entities should they be identified
later.

Relative to the existing API drafts in Notion documents etc., the
specification here opts for minimalism over extra features. This is to
allow iteration over core functionality that is: store, retrieve, check.