Extend blob store to return status and handle /status
Change the blob store interface to separate `Describe` from `Get`, and
extend the content of `blob.Descriptor` to include `blob.Status`.

Reflect the changes on the existing stores. Leave TODOs in place for the
actual information fetching of the deals made by the store to be
implemented in separate PRs.

Change the sever to correctly handle `GET /v0/blob/{id}/status`.