Parse HTTP range headers
Need to parse range headers so that entire range can be requested from singularity. Otherwise it is necessary to read the whole file.

An optimization would be to determine if the ranges are contiguous, and then only make one request that spans contiguous ranges. This is not necessary though if clients already combine contiguous ranges into a single larger range.