feat: add cleanup test, and isolate id mapping code and cleanup code (#218)
Wrote a test for the cleanup code while investigating #201. This
involved moving cleanup code to its own file. I also used this
opportunity to isolate ID mapping code.

The test is passing from the start, so I'm not sure if it'll fix that
issue, but there's a possibility. In any case, it cleans up a bit of
mess.

The cleanup scheduler is a new object with the sole purpose of reading
through the local bin store and removing files based on the result of a
configurable callback, at some interval. A configurable callback allows
it to be mocked easily for the test. In `store.go`, the callback is set
to the `hasDealForAllProviders()` function. This should be tested next.
It'll be a bit more involved since it'll have to be added to the
integration test, since singularity has to be running, but I think I
know how I'll do it.

A few additional functions have been added to local store: `List()` and
`Delete()`. These were previously done inline in `store.go` by manually
working with the store directory files, which was kinda hacky. This puts
that stuff where it belongs.

About the new ID map object: a separation of concerns refactor. Motion
maps blob IDs to Singularity IDs by storing files ending in ".id" in the
same directory, named with the stringified motion ID, and containing the
Singularity ID in the body. This processing was all done inline in
`store.go` With these changes, this behavior is unmodified, so there
should not be any compatibility issues. I originally wrote this because
I thought I would be working with IDs in `cleanup.go`, but my final
solution did not end up needing it. So in the end it is just a general
improvement. It does pave the way for an ID mapping test.

should fix #201, awaiting confirmation by @Angelo-gh3990